+++ src/app/features/lp-optimisers/helpers.ts
