+++ src/app/features/transactions/reducer/transactions/closeTransactionReducer.ts
