+++ src/app/features/transactions/selectors/stateSelector.ts
