+++ src/app/index.ts
@@ -1,5 +1,5 @@
 export * as actions from './features/transactions/actions';
 export * as selectors from './features/transactions/selectors';
 export * from './features/transactions/utilities';
-export { default } from './store';
+export * from './store';
 export * from './types';
