+++ src/app/store.ts
@@ -1,0 +1,32 @@
+import { combineReducers, configureStore, PreloadedState } from '@reduxjs/toolkit';
+import logger from 'redux-logger';
+import createSagaMiddleware from 'redux-saga';
+
+import { aMMsReducer } from './features/aMMs';
+import { lpOptimisersReducer } from './features/lp-optimisers';
+import { transactionsReducer } from './features/transactions/reducer/transactions/transactionsReducer';
+import { saga } from './features/transactions/saga/saga';
+
+const rootReducer = combineReducers({
+  transactions: transactionsReducer,
+  lpOptimisers: lpOptimisersReducer,
+  aMMs: aMMsReducer,
+});
+
+const sagaMiddleware = createSagaMiddleware();
+
+export const setupStore = (preloadedState?: PreloadedState<RootState>) => {
+  const store = configureStore({
+    reducer: rootReducer,
+    middleware: (getDefaultMiddleware) =>
+      getDefaultMiddleware().concat(logger).concat(sagaMiddleware),
+    preloadedState,
+  });
+  sagaMiddleware.run(saga);
+  return store;
+};
+export const store = setupStore();
+
+export type RootState = ReturnType<typeof rootReducer>;
+export type AppStore = ReturnType<typeof setupStore>;
+export type AppDispatch = AppStore['dispatch'];
